Initial commit
I had a working window system with translations earlier, but it wasn't up to my standards of error handling. Next task on the list is to get proper error handling/fallbacks for when something can't be translated (probably have a hardcoded en-US string). After that I need to continue the gfx-hal tutorial so that I can write a pipeline to my standards.

The problem with how I want to write this is that I'm overengineering to make everything as configurable as possible. This means that I have to write a lot of boiler-plate in the beginning. Hopefully it should pay off. I'm going to need a type-checkable scripting language to do what I want though. I'll search crates.io for that, perhaps.